fix gender type

+++ src/center_server/game_database_session.h
@@ -35,7 +35,7 @@ public:
         uint64 guid, 
         const std::string& email,
         const std::string& password,
-        uint8 gender,
+        int32 gender,
         const std::string& nickname,
         const std::string& register_ip,
         uint64 register_timestamp
